% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesnet.R
\name{bayesnet}
\alias{bayesnet}
\title{Solve Bayesian Network}
\usage{
bayesnet(nodes)
}
\arguments{
\item{nodes}{Bayesian Network with Conditional Probability tables (JSON).}
}
\value{
Bayesian Network with posterior distributions.
}
\description{
Solve Bayesian Network
}

