% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{sa}
\alias{sa}
\title{Sensitivity Analysis using Machine Learning}
\usage{
sa(response = NULL, data = NULL, gsd.version = NULL, gsd.dir = "gsd",
  run_gbm = "new", n.trees = 5000, n.samples = NULL, scale_response = F,
  log_response = F, SI_include_level = 5, shrinkage = 0.001,
  train.fraction = 1, bag.fraction = 1, interaction.depth = 10,
  n.minobsinnode = 4, n.extreme = 5, plots = T, response.quan = T,
  response.colors = c("goldenrod", "lightgreen", "lightblue", "lightpink"),
  density.cex = 0.7, pd.cex = 0.7, quants.probs = c(0.5, 0.95, 0.99),
  subset_ind = NULL, endpoints.id = NULL, extremes = FALSE,
  remove.fields = NULL, new.fields = NULL, Rdata.name = NULL,
  label.prefix = "", bright_line = NULL)
}
\arguments{
\item{response}{.}

\item{data}{.}

\item{gsd.version}{.}

\item{gsd.dir}{.}
}
\value{
data.
}
\description{
Sensitivity Analysis using Machine Learning
}

