% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnormcpt.R
\name{tnormcpt}
\alias{tnormcpt}
\title{Solve Truncated Normals for Ranked Nodes}
\usage{
tnormcpt(mu, variance, ntarget, nvpar)
}
\arguments{
\item{mu}{function for mu of the truncated normal distribution.}

\item{ntarget}{number of levels of the target node.}

\item{nVpar}{vector of numbers of levels of each parent.}
}
\value{
conditional probability table.
}
\description{
Solve Truncated Normals for Ranked Nodes
}
\examples{
mu='Staff+Testing'; nVpar=rjson::fromJSON('{"Staff": 5, "Testing": 5}'); ntarget = 5
mu='20*Patrons+10*Reservation'; nVpar=rjson::fromJSON('{"Patrons": 3, "Reservation": 2}'); ntarget = 4
mu='10'; nVpar=0; ntarget = 6; var=1
mu='0.9'; var=0.1; nVpar=1; ntarget = 10;
gisdtR::tnormcpt(mu=.01,var=.1,ntarget=5,nVpar=0)
gisdtR::tnormcpt(mu=mu,var=var,ntarget=ntarget,nVpar=nVpar)
gisdtR::
tnormcpt(mu='0.1*Node53+0.1*Node71',variance='0.1',ntarget=2,nvpar='{"Node53": 3, "Node71": 2}')
SELECT shared.gisdt_tnormcpt('0.1*Node53+0.1*Node71','0.1',2,'{"Node53": 3, "Node71": 2}')
tnormcpt('2.3*Node72+3.7*Node59+5.3*Node74','0.65', 2,'{"Node72": 2, "Node59": 3, "Node74": 2}')
}
